CodeMirror Int: Statically load resource files
codemirror-react's default behavior is to automatically load language
resource files when it is created.  This doesn't work for me as I need
to have my entire bundle prepared ahead of time.

I created a pr to add the ability to disable this behavior
https://github.com/uiwjs/react-codemirror/pull/132

The maintainer was pretty quick to merge my change, so I can statically
load resource files now