Build: Add public path to prod webpack config
This was mistakenly omitted from the prod webpack config when added to
the dev webpack config.

This was causing build issues as resources from nested directories (in
this case /blog) would be searched for in that directory.

For instance, trying to find the bundle (main.js) from a file nested
file (such as /blog/post-name) would result in looking for /blog/main.js
(which 404s) instead of the correct /main.js